pyserial_async/requirements.txt
pyserial-asyncio

